+++ cli/cli/kurtosis_gateway/server/api_container_gateway/api_container_gateway_service_server.go
@@ -303,8 +303,9 @@ func (service *ApiContainerGatewayServiceServer) startRunningConnectionForKurtos
 	localPublicApiPorts := map[string]*kurtosis_core_rpc_api_bindings.Port{}
 	for portId, privateApiPort := range privatePortsFromApi {
 		localPortSpec, found := serviceConnection.GetLocalPorts()[portId]
+		// Skip the private remote port if no public local port is forwarding to it
 		if !found {
-			return nil, stacktrace.NewError("Service requested private port '%v', but no local forwarded port was made for it", portId)
+			continue
 		}
 		localPublicApiPorts[portId] = &kurtosis_core_rpc_api_bindings.Port{
 			Number:   uint32(localPortSpec.GetNumber()),