+++ cli/docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.21.0
+### Breaking Changes
+* Upgraded to engine 1.26.1, core 1.54.1 and container-engine-lib 0.33.0
+  * Users will need to run `kurtosis engine restart`
+
+### Fixes
+* Fixed bug with Gateway erring out when getting service info for service with non TCP protocol port
+
 # 0.20.1
 ### Changes
 * Updated default kubernetes-enclave-data-volume-size to 1024Mb